+++ src/lfx/src/lfx/services/settings/groups/runtime.py
@@ -1,0 +1,54 @@
+from typing import Literal
+
+from pydantic import BaseModel, Field, field_validator
+
+from lfx.log.logger import logger
+
+
+class RuntimeSettings(BaseModel):
+    """Runtime behaviors: event delivery, worker timeouts, polling intervals, public flows, misc toggles.
+
+    Note: ``event_delivery`` is validated here but reads ``workers`` from
+    :class:`ServerSettings`. The composition order in :class:`Settings`
+    guarantees ``workers`` is in ``info.data`` when this validator runs.
+    """
+
+    dev: bool = False
+    """If True, Langflow will run in development mode."""
+
+    event_delivery: Literal["polling", "streaming", "direct"] = "streaming"
+    """How to deliver build events to the frontend. Can be 'polling', 'streaming' or 'direct'."""
+
+    worker_timeout: int = 300
+    """Timeout for the API calls in seconds."""
+
+    public_flow_cleanup_interval: int = Field(default=3600, gt=600)
+    """The interval in seconds at which public temporary flows will be cleaned up.
+    Default is 1 hour (3600 seconds). Minimum is 600 seconds (10 minutes)."""
+    public_flow_expiration: int = Field(default=86400, gt=600)
+    """The time in seconds after which a public temporary flow will be considered expired and eligible for cleanup.
+    Default is 24 hours (86400 seconds). Minimum is 600 seconds (10 minutes)."""
+
+    webhook_polling_interval: int = 0
+    """The polling interval for the webhook in ms. Set to 0 to disable (SSE provides real-time updates)."""
+    fs_flows_polling_interval: int = 10000
+    """The polling interval in milliseconds for synchronizing flows from the file system."""
+
+    health_check_max_retries: int = 5
+    """The maximum number of retries for the health check."""
+
+    max_file_size_upload: int = 1024
+    """The maximum file size for the upload in MB."""
+
+    celery_enabled: bool = False
+
+    @field_validator("event_delivery", mode="before")
+    @classmethod
+    def set_event_delivery(cls, value, info):
+        # If workers > 1, we need to use direct delivery
+        # because polling and streaming are not supported
+        # in multi-worker environments
+        if info.data.get("workers", 1) > 1:
+            logger.warning("Multi-worker environment detected, using direct event delivery")
+            return "direct"
+        return value
